# docs/swagger.yaml
basePath: /
definitions:
...
  serializer.UserJSON:
    properties:
      is_moderator:
        type: boolean
      password:
        type: string
      user_id:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  serializer.UserRespJSON:
    properties:
      is_moderator:
        type: boolean
      user_id:
        type: string
      username:
        type: string
    type: object
  serializer.UserRespWithTokenJSON:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/serializer.UserRespJSON'
    type: object
host: localhost:8080
info:
  contact:
    email: support@insulin.ru
    name: API Support
  description: API для расчета болюсного инсулина
  license:
    name: AS IS (NO WARRANTY)
  title: Insulin Calculation API
  version: "1.0"
paths:
...
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Принимает логин/пароль, возвращает jwt-токен и данные пользователя.
      parameters:
      - description: Логин и пароль
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/serializer.UserJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Ответ с токеном и пользователем
          schema:
            $ref: '#/definitions/serializer.UserRespWithTokenJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход (получение токена)
      tags:
      - users
  /api/users/logout:
    post:
      description: Удаляет токен текущего пользователя из хранилища. Возвращает {"status":"signed_out"}.
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Проблема с получением user_id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка при удалении токена
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выход (удаление токена)
      tags:
      - users
  /api/users/profile:
    get:
      description: Возвращает данные текущего аутентифицированного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/serializer.UserRespJSON'
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить текущего пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет профиль пользователя (может делать только сам пользователь).
      parameters:
      - description: Новые данные профиля
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/serializer.UserJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый профиль
          schema:
            $ref: '#/definitions/serializer.UserRespJSON'
        "400":
          description: Ошибка запроса или авторизации
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещён
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить профиль пользователя
      tags:
      - users
  /api/users/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя. Возвращает URL созданного ресурса
        в Location и тело созданного пользователя.
      parameters:
      - description: Параметры нового пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/serializer.UserJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/serializer.UserRespJSON'
        "400":
          description: Ошибка валидации или входных данных
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - users
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"